services/paragraph-enricher/requirements.txt
pyspark==4.0.0
google-genai
dotenv
confluent-kafka